[FIX] LaunchViewController 이름 변경
